+++ frontend/app/src/background.ts
@@ -1,5 +1,5 @@
 import path from 'path';
-import { app, BrowserWindow, Menu, protocol } from 'electron';
+import { app, BrowserWindow, Menu, MenuItem, protocol } from 'electron';
 import installExtension, { VUEJS3_DEVTOOLS } from 'electron-devtools-installer';
 import windowStateKeeper from 'electron-window-state';
 import { createProtocol } from 'vue-cli-plugin-electron-builder/lib';
@@ -45,6 +45,23 @@ const onReady = async () => {
   ipcSetup(pyHandler, getWindow, closeApp, trayManager, menuActions);
   await createWindow();
   trayManager.listen();
+
+  getWindow().webContents.on('context-menu', (event, props) => {
+    const menu = new Menu();
+    if (props.editFlags.canCut) {
+      menu.append(new MenuItem({ label: 'Cut', role: 'cut' }));
+    }
+
+    if (props.editFlags.canCopy) {
+      menu.append(new MenuItem({ label: 'Copy', role: 'copy' }));
+    }
+
+    if (props.editFlags.canPaste) {
+      menu.append(new MenuItem({ label: 'Paste', role: 'paste' }));
+    }
+
+    menu.popup({ window: getWindow() });
+  });
 };
 
 const lock = app.requestSingleInstanceLock();